fix(interp): Allow return statements in observeField callbacks
This exposed a circular dependency between `brsTypes/`
`interpreter/index.ts`, and `Error.ts`, resolved by extracting
`TypeMismatch.ts` to the `interpreter/` tree since it's currently only
used at BrightScript runtime.

fixes #508
