+++ src/Error.ts
@@ -1,5 +1,4 @@
-import { BrsType, ValueKind } from "./brsTypes";
-import { Location } from "./lexer";
+import type { Location } from "./lexer";
 
 export class BrsError extends Error {
     constructor(message: string, readonly location: Location) {
@@ -43,50 +42,6 @@ export class BrsError extends Error {
     }
 }
 
-/** Wraps up the metadata associated with a type mismatch error. */
-export interface TypeMismatchMetadata {
-    /**
-     * The base message to use for this error. Should be as helpful as possible, e.g.
-     * "Attempting to subtract non-numeric values".
-     */
-    message: string;
-    /** The value on the left-hand side of a binary operator, or the *only* value for a unary operator. */
-    left: TypeAndLocation;
-    /** The value on the right-hand side of a binary operator. */
-    right?: TypeAndLocation;
-}
-
-export type TypeAndLocation = {
-    /** The type of a value involved in a type mismatch. */
-    type: BrsType | ValueKind;
-    /** The location at which the offending value was resolved. */
-    location: Location;
-};
-
-/**
- * Creates a "type mismatch"-like error message, but with the appropriate types specified.
- * @return a type mismatch error that will be tracked by this module.
- */
-export class TypeMismatch extends BrsError {
-    constructor(mismatchMetadata: TypeMismatchMetadata) {
-        let messageLines = [
-            mismatchMetadata.message,
-            `    left: ${ValueKind.toString(getKind(mismatchMetadata.left.type))}`,
-        ];
-        let location = mismatchMetadata.left.location;
-
-        if (mismatchMetadata.right) {
-            messageLines.push(
-                `    right: ${ValueKind.toString(getKind(mismatchMetadata.right.type))}`
-            );
-
-            location.end = mismatchMetadata.right.location.end;
-        }
-
-        super(messageLines.join("\n"), location);
-    }
-}
-
 /**
  * Logs a detected BRS error to console.
  * @param err the error to log to console
@@ -102,17 +57,4 @@ export function logConsoleError(err: BrsError) {
  */
 export function getLoggerUsing(errorStream: NodeJS.WriteStream): (err: BrsError) => boolean {
     return (err) => errorStream.write(err.format());
-}
-
-/**
- * Returns the `.kind` property of a `BrsType`, otherwise returns the provided `ValueKind`.
- * @param maybeType the `BrsType` to extract a `.kind` field from, or the `ValueKind` to return directly
- * @returns the `ValueKind` for `maybeType`
- */
-function getKind(maybeType: BrsType | ValueKind): ValueKind {
-    if (typeof maybeType === "number") {
-        return maybeType;
-    } else {
-        return maybeType.kind;
-    }
 }
